+++ lib/libesp32/berry/default/be_modtab.c
@@ -187,7 +187,10 @@ be_extern_native_class(lv_wifi_bars_icon);
 be_extern_native_class(lv_signal_arcs);
 be_extern_native_class(lv_wifi_arcs);
 be_extern_native_class(lv_wifi_arcs_icon);
+be_extern_native_class(lv_clock);
 be_extern_native_class(lv_clock_icon);
+
+be_extern_native_class(int64);
 
 BERRY_LOCAL bclass_array be_class_table = {
     /* first list are direct classes */
@@ -231,6 +234,7 @@ BERRY_LOCAL bclass_array be_class_table = {
     &be_native_class(lv_signal_arcs),
     &be_native_class(lv_wifi_arcs),
     &be_native_class(lv_wifi_arcs_icon),
+    &be_native_class(lv_clock),
     &be_native_class(lv_clock_icon),
 #endif // USE_LVGL
 
@@ -246,6 +250,9 @@ BERRY_LOCAL bclass_array be_class_table = {
 #endif // USE_UFILESYS
     &be_native_class(AudioOpusDecoder),
 #endif // USE_I2S_AUDIO_BERRY
+#ifdef USE_BERRY_INT64
+    &be_native_class(int64),
+#endif
 #if defined(USE_MI_ESP32) && !defined(USE_BLE_ESP32)
     &be_native_class(MI32),
     &be_native_class(BLE),