+++ tasmota/html_uncompressed/HTTP_SCRIPT_TEMPLATE.h
@@ -1,7 +1,7 @@
 const char HTTP_SCRIPT_TEMPLATE[] PROGMEM =
   "function ld(u,f){"
     "var x=new XMLHttpRequest();"
-    "x.onreadystatechange=function(){"
+    "x.onreadystatechange=()=>{"
       "if(this.readyState==4&&this.status==200){"
         "f(this);"
       "}"