+++ tasmota/language/zh_CN.h
@@ -77,6 +77,7 @@
 #define D_COUNTER "计数器"
 #define D_CT_POWER "CT Power"
 #define D_CURRENT "电流"          // As in Voltage and Current
+#define D_CURRENT_NEUTRAL "Current Neutral"
 #define D_DATA "数据:"
 #define D_DARKLIGHT "暗"
 #define D_DEBUG "调试"